fix: rename chat package
